Implemented AnalysisBoard
